[Demo Comments and little changes]
cleaned up the code of the python file for talking
cleaned up M1 Demo Code and made some little changes
